removed the testbook protocol handling from kernel
